attendance modularised
